created ClickableAvatarList, refactored, ItemTable visually correct
